AP_IOMCU: moved protocol structures to a common header
