+++ libraries/AP_IOMCU/AP_IOMCU.cpp
@@ -17,102 +17,6 @@
 
 extern const AP_HAL::HAL &hal;
 
-#define PKT_MAX_REGS 32
-
-//#define IOMCU_DEBUG
-
-struct PACKED IOPacket {
-	uint8_t 	count:6;
-	uint8_t 	code:2;
-	uint8_t 	crc;
-	uint8_t 	page;
-	uint8_t 	offset;
-	uint16_t	regs[PKT_MAX_REGS];
-
-    // get packet size in bytes
-    uint8_t get_size(void) const {
-        return count*2 + 4;
-    }
-};
-
-/*
-  values for pkt.code
- */
-enum iocode {
-    // read types
-    CODE_READ = 0,
-    CODE_WRITE = 1,
-
-    // reply codes
-    CODE_SUCCESS = 0,
-    CODE_CORRUPT = 1,
-    CODE_ERROR = 2
-};
-
-// IO pages
-enum iopage {
-    PAGE_CONFIG = 0,
-    PAGE_STATUS = 1,
-    PAGE_ACTUATORS = 2,
-    PAGE_SERVOS = 3,
-    PAGE_RAW_RCIN = 4,
-    PAGE_RCIN = 5,
-    PAGE_RAW_ADC = 6,
-    PAGE_PWM_INFO = 7,
-    PAGE_SETUP = 50,
-    PAGE_DIRECT_PWM = 54,
-    PAGE_FAILSAFE_PWM = 55,
-    PAGE_DISARMED_PWM = 108,
-};
-
-// pending IO events to send, used as an event mask
-enum ioevents {
-    IOEVENT_INIT=1,
-    IOEVENT_SEND_PWM_OUT,
-    IOEVENT_SET_DISARMED_PWM,
-    IOEVENT_SET_FAILSAFE_PWM,
-    IOEVENT_FORCE_SAFETY_OFF,
-    IOEVENT_FORCE_SAFETY_ON,
-    IOEVENT_SET_ONESHOT_ON,
-    IOEVENT_SET_RATES,
-    IOEVENT_GET_RCIN,
-    IOEVENT_ENABLE_SBUS,
-    IOEVENT_SET_HEATER_TARGET,
-    IOEVENT_SET_DEFAULT_RATE,
-    IOEVENT_SET_SAFETY_MASK,
-};
-
-// setup page registers
-#define PAGE_REG_SETUP_FEATURES	0
-#define P_SETUP_FEATURES_SBUS1_OUT	1
-#define P_SETUP_FEATURES_SBUS2_OUT	2
-#define P_SETUP_FEATURES_PWM_RSSI   4
-#define P_SETUP_FEATURES_ADC_RSSI   8
-#define P_SETUP_FEATURES_ONESHOT   16
-
-#define PAGE_REG_SETUP_ARMING 1
-#define P_SETUP_ARMING_IO_ARM_OK (1<<0)
-#define P_SETUP_ARMING_FMU_ARMED (1<<1)
-#define P_SETUP_ARMING_RC_HANDLING_DISABLED (1<<6)
-#define P_SETUP_ARMING_SAFETY_DISABLE_ON	(1 << 11) // disable use of safety button for safety off->on
-#define P_SETUP_ARMING_SAFETY_DISABLE_OFF	(1 << 12) // disable use of safety button for safety on->off
-
-#define PAGE_REG_SETUP_PWM_RATE_MASK 2
-#define PAGE_REG_SETUP_DEFAULTRATE   3
-#define PAGE_REG_SETUP_ALTRATE       4
-#define PAGE_REG_SETUP_REBOOT_BL    10
-#define PAGE_REG_SETUP_CRC			11
-#define PAGE_REG_SETUP_SBUS_RATE    19
-#define PAGE_REG_SETUP_IGNORE_SAFETY 20 /* bitmask of surfaces to ignore the safety status */
-#define PAGE_REG_SETUP_HEATER_DUTY_CYCLE 21
-
-// magic value for rebooting to bootloader
-#define REBOOT_BL_MAGIC 14662
-
-#define PAGE_REG_SETUP_FORCE_SAFETY_OFF 12
-#define PAGE_REG_SETUP_FORCE_SAFETY_ON  14
-#define FORCE_SAFETY_MAGIC 22027
-
 AP_IOMCU::AP_IOMCU(AP_HAL::UARTDriver &_uart) :
     uart(_uart)
 {}
